Push all functions into the PINGData class.
