+++ similarity.py
@@ -12,9 +12,7 @@ from matplotlib import pyplot as plt
 from ping.analysis.similarity import (compare_similarity_vectors,
                                       compute_similarity_vectors,
                                       visualize_similarity_matrices)
-from ping.data import (PINGData, which_hemi, get_nonhemi_key, is_nonimaging_key, 
-                       get_anatomical_name, anatomical_sort,
-                       prefix2text, get_measure_key)
+from ping.data import PINGData
 from ping.utils import filter_dict
 from research.asymmetry import is_ai_key
 from research.data import get_all_data
@@ -45,19 +43,19 @@ else:
     prefix = sys.argv[1].split(',')
     prefix_filter_fn = lambda k, v: np.any([k.startswith(p) for p in prefix])
 
-    # Get metric
-    metric = 'partial-correlation' if len(sys.argv) <= 2 else sys.argv[2]
-    key_locations = list(filt_fns.keys()) if len(sys.argv) <= 3 else sys.argv[3].split(',')
-
     # Determine filters for selecting the similarity groupings. 
     filt_fns = OrderedDict((
-        ('Asymmetry Index', lambda key: (is_ai_key(key) and np.all([substr not in key for substr in ['_TOTAL', 'LH_PLUS_RH']])) or is_nonimaging_key(key)),
-        ('Left Hemisphere', lambda key: which_hemi(key) == 'lh' or is_nonimaging_key(key)),
-        ('Right Hemisphere', lambda key: which_hemi(key) == 'rh' or is_nonimaging_key(key)),
+        ('Asymmetry Index', lambda key: (is_ai_key(key) and np.all([substr not in key for substr in ['_TOTAL', 'LH_PLUS_RH']])) or PINGData.is_nonimaging_key(key)),
+        ('Left Hemisphere', lambda key: PINGData.which_hemi(key) == 'lh' or PINGData.is_nonimaging_key(key)),
+        ('Right Hemisphere', lambda key: PINGData.which_hemi(key) == 'rh' or PINGData.is_nonimaging_key(key)),
         ('all', lambda key: True)))
     filt_fns['LHAI'] = partial(lambda key, f1, f2: f1(key) or f2(key),
                                f1=filt_fns['Asymmetry Index'],
                                f2=filt_fns['Left Hemisphere'])
+
+    # Get metric
+    metric = 'partial-correlation' if len(sys.argv) <= 2 else sys.argv[2]
+    key_locations = list(filt_fns.keys()) if len(sys.argv) <= 3 else sys.argv[3].split(',')
 
     # Get the key locations
     filt_fns = OrderedDict(((k, filt_fns[k]) for k in key_locations))
@@ -69,7 +67,7 @@ else:
     # Do the similarity computation; order by anatomy.
     sim_dict, sim_keys = compute_similarity_vectors(p_data.data_dict,
                                                     filt_fns=filt_fns,
-                                                    sort_fn=anatomical_sort,
+                                                    sort_fn=PINGData.anatomical_sort,
                                                     metric=metric)
 
     # Split the keys into the class (the prefix)
@@ -78,19 +76,19 @@ else:
     labels = []
     class_labels = []
     for ki, key in enumerate(good_keys):
-        class_label, label = [(p, get_nonhemi_key(key)[(len(p) + 1):])
+        class_label, label = [(p, PINGData.get_nonhemi_key(key)[(len(p) + 1):])
                               for p in prefix
                               if key.startswith(p)][0]
         class_labels.append(class_label)
-        labels.append(get_anatomical_name(label))
+        labels.append(PINGData.get_anatomical_name(label))
 
     # Compare matrices (printed)
     compare_similarity_vectors(sim_dict)
-    
+
     #  Display the similarity matrices.
     ax = visualize_similarity_matrices(sim_dict, labels=labels,
                                        class_labels=class_labels)
-    # ax.get_figure().suptitle(', '.join([prefix2text(p) for p in prefix]), fontsize=24)
+    # ax.get_figure().suptitle(', '.join([PINGData.prefix2text(p) for p in prefix]), fontsize=24)
 
     for key, mat in sim_dict.items():
         import scipy.spatial
